requirements.txt
msgpack==1.0.8
packaging==24.1
paho-mqtt==2.1.0
PyYAML==6.0.2
typing_extensions==4.12.2
wrapt==1.16.0